BuildCon  mini projects
Buildcon has data models for all types of automated reports, benchmarking and predictions. As a construction company manager, it’s not your job to build the data models and deal with analytics - Buildcon will do it for you. It will provide reports which are customizable and adjustable, that way the application tailored specifically for your needs.